linux-custom: honor KERNEL_FILE setting
vmlinuz was hard-coded in the postinst script generated by the
build-kernel.sh script. This really needs to be ${KERNEL_FILE}
to support architectures where a vmlinux file is used instead
(mipsel in particular). Rename the "vmlinuz" local variable to
kernel_file so that it is clear it is derived from KERNEL_FILE.

As we now have multiple recipes using KERNEL_FILE (image and
kernel recipes), its default setting (vmlinuz) was moved to
bitbake.conf (defaults to vmlinuz for all architectures but
mipsel where it is set to vmlinux)

Signed-off-by: Cedric Hombourger <Cedric_Hombourger@mentor.com>